+++ app/models/claim/geckoboard_dataset.rb
@@ -13,12 +13,12 @@ class Claim
       Geckoboard::StringField.new(:policy, name: "Policy"),
       Geckoboard::DateTimeField.new(:submitted_at, name: "Submitted at"),
       Geckoboard::StringField.new(:passed_checking_deadline, name: "Passed checking deadline"),
-      Geckoboard::StringField.new(:checked, name: "Checked"),
+      Geckoboard::StringField.new(:check, name: "Check"),
       Geckoboard::DateTimeField.new(:checked_at, name: "Checked at"),
-      Geckoboard::StringField.new(:check_result, name: "Check result"),
       Geckoboard::NumberField.new(:number_of_days_to_check, name: "Number of days to check", optional: true),
       Geckoboard::StringField.new(:paid, name: "Paid"),
       Geckoboard::DateTimeField.new(:paid_at, name: "Paid at"),
+      Geckoboard::MoneyField.new(:award_amount, name: "Award amount", optional: true, currency_code: "GBP"),
     ]
     BATCH_SIZE = 500
 
@@ -36,8 +36,6 @@ class Claim
       client.datasets.delete(dataset_name)
     end
 
-    private
-
     def data_for_claims(claims)
       claims.map do |claim|
         {
@@ -45,15 +43,17 @@ class Claim
           policy: claim.policy.to_s,
           submitted_at: claim.submitted_at,
           passed_checking_deadline: claim.check_deadline_date.past?.to_s,
-          checked: claim.check.present?.to_s,
+          check: claim.check.present? ? claim.check.result : "unchecked",
           checked_at: claim.check.present? ? claim.check.created_at : placeholder_date_for_nil_value,
-          check_result: claim.check.present? ? claim.check.result : "",
           number_of_days_to_check: claim.check&.number_of_days_since_claim_submitted,
           paid: claim.scheduled_payment_date.present?.to_s,
           paid_at: claim.scheduled_payment_date.presence || placeholder_date_for_nil_value,
+          award_amount: claim.eligibility.present? ? (claim.award_amount * 100).to_i : nil, # Geckoboard expects currency as an integer of pence - see https://api-docs.geckoboard.com/#money-format
         }
       end
     end
+
+    private
 
     def client
       @client ||= Geckoboard.client(ENV.fetch("GECKOBOARD_API_KEY"))